Pull the `Utils` methods into `TestFiles`.
Also start to get the docs to represent the intended final arrangement.
