+++ local-modules/testing-server/README.md
@@ -1,11 +1,11 @@
 testing-server
 ==============
 
-This module serves dual (but related) purposes:
+This module serves multiple related purposes:
 
-* Contains general code for identifying test files within a subproduct (client
-  or server).
-* Contains the code needed to run server tests.
+* Identifying test files within a subproduct (client or server).
+* Directly running server tests.
+* Arm's-length running of client tests via a captive browser.
 
 A bit of subtlety here is that it is code on the server side which prepares
 the client-side testing code. This module helps with that. In that regard, it
